pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "benchwarmer"
version = "0.1.0"
description = "Algorithm benchmarking with NL intake and Perplexity-backed search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyMuPDF>=1.24.0",
    "anthropic>=0.39.0",
    "modal>=0.64.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["benchwarmer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
